eduz_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
python-dotenv
alembic
pydantic[email]
python-multipart
pytest
httpx